create job file for docid and fix runtime filenames
