+++ transforms/universal/doc_id/Dockerfile.python
@@ -1,39 +1,48 @@
-FROM docker.io/python:3.10.14-slim-bullseye
+FROM docker.io/python:3.12-slim
 
 RUN pip install --upgrade --no-cache-dir pip 
 
 # install pytest
 RUN pip install --no-cache-dir pytest
 
+
 # Create a user and use it to run the transform
-RUN useradd -ms /bin/bash dpk
-USER dpk
+RUN useradd -d /home/dpk -ms /bin/bash -g root dpk
 WORKDIR /home/dpk
+USER dpk
+
 ARG DPK_WHEEL_FILE_NAME
 ARG TRANSFORM_NAME
 
 # Copy and install data processing libraries 
 # These are expected to be placed in the docker context before this is run (see the make image).
-COPY --chown=dpk:users data-processing-dist data-processing-dist
+COPY --chown=dpk:root data-processing-dist data-processing-dist
 RUN  pip install data-processing-dist/${DPK_WHEEL_FILE_NAME}
 
 # END OF STEPS destined for a data-prep-kit base image 
 
-COPY --chown=dpk:users dpk_${TRANSFORM_NAME}/ dpk_${TRANSFORM_NAME}/
-COPY --chown=dpk:users requirements.txt requirements.txt 
+COPY --chown=dpk:root dpk_${TRANSFORM_NAME}/ dpk_${TRANSFORM_NAME}/
+COPY --chown=dpk:root requirements.txt requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Set environment
-ENV PYTHONPATH /home/dpk
+ENV PYTHONPATH=/home/dpk
+ENV HOME=/home/dpk
+
+USER root
+RUN chown -R dpk:root /home/dpk && \
+   chmod -R g=u /home/dpk
+USER dpk
+
 
 # Put these at the end since they seem to upset the docker cache.
 ARG BUILD_DATE
 ARG GIT_COMMIT
 ARG TRANSFORM_PATH
 ARG GIT_URL
-ENV GIT_URL $GIT_URL
-ENV TRANSFORM_PATH $TRANSFORM_PATH
-ENV BUILD_DATE $BUILD_DATE
-ENV GIT_COMMIT $GIT_COMMIT
+ENV GIT_URL=$GIT_URL
+ENV TRANSFORM_PATH=$TRANSFORM_PATH
+ENV BUILD_DATE=$BUILD_DATE
+ENV GIT_COMMIT=$GIT_COMMIT
 LABEL build-date=$BUILD_DATE
 LABEL git-commit=$GIT_COMMIT
