+++ transforms/universal/doc_id/Makefile
@@ -13,33 +13,39 @@ TRANSFORM_NAME=$(shell basename `pwd`)
 
 ################################################################################
 
+# Developers should adapt this according to their implementation if they do not follow the proposed template
+# We continue to refine this. The current settings are based on current template
+TRANSFORM_PYTHON_SRC="-m dpk_$(TRANSFORM_NAME).runtime"
+TRANSFORM_RAY_SRC="-m dpk_$(TRANSFORM_NAME).ray.runtime"
+TRANSFORM_SPARK_SRC="-m dpk_$(TRANSFORM_NAME).spark.runtime"
+
 
 
 run-cli-python-sample: 
 	make venv
 	source venv/bin/activate && \
-	$(PYTHON) -m dpk_$(TRANSFORM_NAME).transform_python \
+	$(PYTHON) -m dpk_$(TRANSFORM_NAME).runtime \
                 --data_local_config "{ 'input_folder' : 'test-data/input', 'output_folder' : 'output'}"  \
                 --doc_id_int True
 
 run-cli-python-s3: 
 	make venv
 	source venv/bin/activate && \
-	$(PYTHON) -m dpk_$(TRANSFORM_NAME).transform_python \
+	$(PYTHON) -m dpk_$(TRANSFORM_NAME).runtime \
                 --data_s3_config "{ 'input_folder' : 'cloud-object-storage-cos-standard-k1l/doc-id-test-data/input', 'output_folder' : 'cloud-object-storage-cos-standard-k1l/doc-id-output'}"  \
                 --doc_id_int True
 
 run-cli-spark-sample: 
 	make venv
 	source venv/bin/activate && \
-	$(PYTHON) -m dpk_$(TRANSFORM_NAME).spark.transform \
+	$(PYTHON) -m dpk_$(TRANSFORM_NAME).spark.runtime \
                 --data_local_config "{ 'input_folder' : 'test-data/input', 'output_folder' : 'output'}"  \
                 --doc_id_int True
 
 run-cli-ray-sample: 
 	make venv
 	source venv/bin/activate && \
-	$(PYTHON) -m dpk_$(TRANSFORM_NAME).ray.transform \
+	$(PYTHON) -m dpk_$(TRANSFORM_NAME).ray.runtime \
                 --run_locally True --data_local_config "{ 'input_folder' : 'test-data/input', 'output_folder' : 'output'}"  \
                 --doc_id_int True
 
@@ -48,7 +54,7 @@ run-docker-python-sample:
 	make DOCKER_PLATFORM=linux/arm64 image-python
 	mkdir -p ./output
 	$(DOCKER) run -v ./test-data:/test-data -v ./output:/output -it localhost/doc_id-python \
-                python -m dpk_$(TRANSFORM_NAME).transform_python \
+                python -m dpk_$(TRANSFORM_NAME).runtime \
                 --data_local_config "{ 'input_folder' : '/test-data/input', 'output_folder' : '/output'}"  \
                 --doc_id_int True
