+++ transforms/universal/doc_id/dpk_doc_id/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
 from .transform import *
 from .local_python import *
-from .transform_python import *
+from .runtime import *
 from .local import *
