+++ transforms/universal/doc_id/kfp_ray/doc_id_wf.py
@@ -29,7 +29,7 @@ task_image = "quay.io/dataprep1/data-prep-kit/doc_id-ray:latest"
 S3_SECRET = "s3-secret"
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "-m dpk_doc_id.ray.transform"
+EXEC_SCRIPT_NAME: str = "-m dpk_doc_id.ray.runtime"
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:latest"
