+++ transforms/universal/doc_id/test/test_doc_id_python.py
@@ -23,7 +23,7 @@ from dpk_doc_id.transform import (
     int_column_name_cli_param,
     start_id_cli_param,
 )
-from dpk_doc_id.transform_python import DocIDPythonTransformRuntimeConfiguration
+from dpk_doc_id.runtime import DocIDPythonTransformRuntimeConfiguration
 
 
 class TestPythonDocIDTransform(AbstractTransformLauncherTest):
